requirements.txt
openai
google-geneai
streamlit
python-dotenv
